[feat] 신고처리 API 구현

+++ src/main/java/com/ducks/goodsduck/admin/service/ItemService.java
@@ -3,13 +3,23 @@ package com.ducks.goodsduck.admin.service;
 import com.ducks.goodsduck.admin.model.entity.*;
 import com.ducks.goodsduck.admin.model.entity.Image.Image;
 import com.ducks.goodsduck.admin.model.entity.Image.ItemImage;
+import com.ducks.goodsduck.admin.model.enums.ImageType;
 import com.ducks.goodsduck.admin.repository.ChatRepository;
 import com.ducks.goodsduck.admin.repository.ItemRepository;
+import com.ducks.goodsduck.admin.repository.PostRepository;
+import com.ducks.goodsduck.admin.repository.UserRepository;
+import com.ducks.goodsduck.admin.repository.image.ImageRepository;
+import com.ducks.goodsduck.admin.repository.image.PostImageRepository;
+import com.ducks.goodsduck.admin.repository.pricepropose.PriceProposeRepositoryCustom;
+import com.ducks.goodsduck.admin.repository.review.ReviewRepositoryCustom;
 import com.ducks.goodsduck.admin.repository.userchat.UserChatRepository;
 import com.ducks.goodsduck.admin.repository.image.ItemImageRepository;
 import com.ducks.goodsduck.admin.repository.pricepropose.PriceProposeRepository;
 import com.ducks.goodsduck.admin.repository.review.ReviewRepository;
+import com.ducks.goodsduck.admin.repository.userchat.UserChatRepositoryCustom;
 import com.ducks.goodsduck.admin.repository.useritem.UserItemRepository;
+import com.ducks.goodsduck.admin.repository.useritem.UserItemRepositoryCustom;
+import com.ducks.goodsduck.admin.util.PropertyUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -25,34 +35,37 @@ import java.util.List;
 @Slf4j
 public class ItemService {
 
+    String accessKeySecretManager = PropertyUtil.getProperty("cloud.aws.credentials.accessKeySecretManager");
+    String secretKeySecretManager = PropertyUtil.getProperty("cloud.aws.credentials.secretKeySecretManager");
+
     private final ItemRepository itemRepository;
-    private final ItemImageRepository itemImageRepository;
-    private final PriceProposeRepository priceProposeRepository;
-    private final UserItemRepository userItemRepository;
     private final UserChatRepository userChatRepository;
+    private final UserItemRepository userItemRepository;
+    private final PriceProposeRepository priceProposeRepository;
+    private final ItemImageRepository itemImageRepository;
     private final ChatRepository chatRepository;
     private final ReviewRepository reviewRepository;
+
+    private final ImageUploadService imageUploadService;
 
     public Long delete(Long itemId) {
 
         try {
             Item deleteItem = itemRepository.findById(itemId).get();
 
-            // user's item 목록 삭제
-            User user = deleteItem.getUser();
-            List<Item> itemsOfUser = user.getItems();
-            itemsOfUser.remove(deleteItem);
-
             // image 연관 삭제
             List<ItemImage> deleteImages = deleteItem.getImages();
+            for (ItemImage deleteImage : deleteImages) {
+                imageUploadService.deleteImage(deleteImage, ImageType.ITEM);
+            }
             itemImageRepository.deleteAllInBatch(deleteImages);
 
             // pricePropose 연관 삭제
-            List<PricePropose> deletePriceProposes = priceProposeRepository.findAllByItemIdWithAllStatus(itemId);
+            List<PricePropose> deletePriceProposes = priceProposeRepository.findAllByItemIdWithAllStatus(deleteItem.getId());
             priceProposeRepository.deleteAllInBatch(deletePriceProposes);
 
             // userChat 연관 삭제
-            List<UserChat> deleteUserChats = userChatRepository.findByItemId(itemId);
+            List<UserChat> deleteUserChats = userChatRepository.findByItemId(deleteItem.getId());
             userChatRepository.deleteAllInBatch(deleteUserChats);
 
             // chat 삭제
@@ -62,15 +75,14 @@ public class ItemService {
             }
             chatRepository.deleteAllInBatch(deleteChats);
 
-            // TODO
-            // review 연관 삭제 (삭제용 데이터 0번 아이템으로 교체)
-            List<Review> deleteItemOfReviews = reviewRepository.findByItemId(itemId);
+            // review 연관 삭제
+            List<Review> deleteItemOfReviews = reviewRepository.findByItemId(deleteItem.getId());
             for (Review deleteItemOfReview : deleteItemOfReviews) {
-                deleteItemOfReview.deleteItem();
+                deleteItemOfReview.setItem(null);
             }
 
             // userItem 연관 삭제
-            List<UserItem> deleteUserItems = userItemRepository.findByItemId(itemId);
+            List<UserItem> deleteUserItems = userItemRepository.findByItemId(deleteItem.getId());
             userItemRepository.deleteAllInBatch(deleteUserItems);
 
             // item 삭제
@@ -85,12 +97,10 @@ public class ItemService {
     public Long deleteV2(Long itemId) {
 
         try {
-            Item deleteItem = itemRepository.findById(itemId).get();
-
-            // user's Item 목록 삭제
-            User user = deleteItem.getUser();
-            List<Item> deleteItemsOfUser = user.getItems();
-            deleteItemsOfUser.remove(deleteItem);
+            Item deleteItem = itemRepository.findById(itemId).orElseGet(null);
+            if(deleteItem == null) {
+                return -1L;
+            }
 
             // image 연관 삭제
             List<ItemImage> deleteItemImages = deleteItem.getImages();
